DEV-12829 Further refinement to make more generic
