+++ index.js
@@ -1,11 +1,5 @@
-import { isHealthUrl, isResourceUrl } from './helpers/url-string'
-import VmdBaseLogger from './logging/vmd-base-logger'
-import VmdHapiLogger from './logging/vmd-hapi-logger'
+import VmdHapiLogger from './logging/vmd-logger'
 
 export {
-    // Ideally these would be exposed via a separate library to be shared
-    isHealthUrl, isResourceUrl,
-
-    VmdBaseLogger,
     VmdHapiLogger
 }
